+++ backend/src/models/Location.js
@@ -28,7 +28,7 @@ class Location extends Model {
         }
       }
     }
-    }
+  }
 }
 
 module.exports = Location;
